refactor(core/utils): convert to ES6 module
