+++ src/core/remove-respec.js
@@ -1,8 +1,8 @@
 // Module core/remove-respec
 // Removes all ReSpec artefacts right before processing ends
-import utils from "core/utils";
+import { removeReSpec } from "core/utils";
 import { sub } from "core/pubsubhub";
 
 sub('end-all', () => {
-  utils.removeReSpec(document);
+  removeReSpec(document);
 }, { once: true });
