+++ src/core/worker.js
@@ -6,7 +6,7 @@
  */
 
 // Opportunistically preload syntax highlighter, which is used by the worker
-import utils from "core/utils";
+import { createResourceHint } from "core/utils";
 import workerScript from "deps/text!../../worker/respec-worker.js";
 // Opportunistically preload syntax highlighter
 const hint = {
@@ -14,7 +14,7 @@ const hint = {
   href: "https://www.w3.org/Tools/respec/respec-highlight.js",
   as: "script",
 };
-const link = utils.createResourceHint(hint);
+const link = createResourceHint(hint);
 document.head.appendChild(link);
 
 const workerURL = URL.createObjectURL(new Blob([workerScript], {type : 'application/javascript'}));
